added: loadresource TU
- Refactored initialiseContinueBoardArray(). Many inner-called functions
have been moved from "game"'s TU to the new "loadresource" TU.
- The functions moved are wrapped in the Game::Loader namespace. Helper
functions for the moved functions are in the TU's anon namespace.
- load_game_stats_from_file() with its helper functions now return a
std::tuple containing a load_ok flag and a special tuple for the
variables "gameboard.score" and "gameboard.moveCount".
